backend/pyproject.toml
[project]
name = "stt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.0",
    "google-cloud-speech>=2.34.0",
    "google-genai>=0.3.0",
    "websockets>=12.0",
    "uvicorn>=0.38.0",
    "pydantic-settings>=2.11.0",
]
